Initial IconToggleButton Tests
