+++ pubspec.yaml
@@ -17,6 +17,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.0
+  test: ^1.21.4
 
 flutter:
   uses-material-design: true
